requirements.txt
Flask==3.0.3
flask_wtf==1.2.1
pillow==10.4.0
pytest==8.3.2
google-cloud-logging